minor fixes to NewsCard component
