+++ components/newsletter/NewsCard.jsx
@@ -1,6 +1,5 @@
 import React from "react";
 import styles from "@/styles/newsletter/NewsletterCard.module.css";
-import newsContent from "@/data/news.json";
 
 const NewsletterCard = ({ image, title, description, link, onCardClick }) => {
   const handleCardClick = () => {
@@ -23,39 +22,4 @@ const NewsletterCard = ({ image, title, description, link, onCardClick }) => {
   );
 };
 
-const NewsletterCardGallery = () => {
-  const handleCustomCardClick = () => {
-    console.log("Custom click handler for the card");
-  };
-
-  return (
-    <div className={styles.container}>
-      <div className={styles.heading}>
-        <p>
-          Check out our other <strong>Newsletters</strong>
-        </p>
-      </div>
-      <div className={styles.subheading}>
-        <p>
-          Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do
-          eiusmod tempor incididunt ut labore et dolore magna aliqua.
-        </p>
-      </div>
-      <div className={styles.grid}>
-        {newsContent.map((news, idx) => (
-          <div key={news.id}>
-            <NewsletterCard
-              image={news.src}
-              title={news.title}
-              description={news.description.slice(0, 100) + "..."}
-              link={news.link}
-              onCardClick={handleCustomCardClick}
-            />
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-};
-
-export default NewsletterCardGallery;
+export default NewsletterCard;
