+++ styles/newsletter/News.module.css
@@ -23,3 +23,59 @@
   gap: 2rem;
   padding: 1rem 1rem 1rem 1rem;
 }
+
+.container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  margin-top: 3rem;
+  width: 100%;
+  padding: 1rem 1rem 1rem 1rem;
+}
+
+.heading {
+  color: #eee;
+  margin-bottom: 1rem;
+  font-family: Cabin;
+  font-weight: 400;
+  font-style: Regular;
+  font-size: 36px;
+  line-height: 100%;
+  letter-spacing: 0%;
+  vertical-align: middle;
+  text-align: center;
+}
+
+.subheading {
+  color: #ccc;
+  margin-bottom: 0.5rem;
+  font-family: Cabin;
+  font-weight: 400;
+  font-style: Regular;
+  font-size: 18px;
+  line-height: 100%;
+  letter-spacing: 0%;
+  vertical-align: middle;
+}
+
+.grid {
+  display: grid;
+  width: 65%;
+  height: auto;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 1rem;
+}
+
+@media (max-width: 480px) {
+  .heading {
+    font-size: 28px;
+  }
+  .subheading {
+    display: none;
+  }
+  .grid {
+    width: 90%;
+    grid-template-columns: repeat(1, 1fr);
+  }
+}
